pyproject.toml
[project]
name = "torch1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "tensorboard>=2.19.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]
